보안성 검사

+++ ewp-api-agent/src/main/java/kr/co/ewp/api/controller/LogManageController.java
@@ -17,51 +17,59 @@ public class LogManageController {
 		prettyLog.title("로그파일 관리");
 		Date today = new Date();
 		
-		String realPath = LogManageController.class.getResource("").getPath();
+		String realPath = "";
+		if(LogManageController.class.getResource("") != null){
+			realPath = LogManageController.class.getResource("").getPath();
+		}
 		prettyLog.append("realPath  ", realPath);
 		String path[] = realPath.split("ewpsp_batch");
-		String _realPath = path[0].replace("file:", "")+"ewpsp_batch/logs";
+		String _realPath = path[0].replace("file:", "")+"ewpsp_batch"+File.separator+"logs";
 		prettyLog.append("_realPath  ", _realPath);
 		
-		File dirFile = new File(realPath);
+		File dirFile = new File(_realPath);
 		File [] fileList = dirFile.listFiles();
 		String substringTxt = "ewp-api.";
 		int deleteCnt = 0;
-		for(File tempFile : fileList) {
-			if(tempFile.isFile()) {
-				String tempFileName = tempFile.getName();
-				prettyLog.append("FileName=", tempFileName);
-				
-				int extIdx = tempFileName.lastIndexOf(".");
-				String _tempFileName = tempFileName.substring(0, extIdx);
-				String __tempFileName = _tempFileName.replaceAll(substringTxt, "");
-				
-				int logDate_yyyy = Integer.parseInt(__tempFileName.substring(0, 4));
-				int logDate_MM = Integer.parseInt(__tempFileName.substring(4, 6));
-				int logDate_dd = Integer.parseInt(__tempFileName.substring(6, 8));
-				
-				Calendar cal = Calendar.getInstance();
-				cal.set(logDate_yyyy, logDate_MM-1, logDate_dd);
-				cal.set(Calendar.MILLISECOND, 0);
-				Date logDate = cal.getTime();
-				
-				long calDate = today.getTime()-logDate.getTime();
-				long calDateDays = calDate / ( 24*60*60*1000 ); 
-				
-				if(calDateDays > 31) {
-					tempFile.delete();
-					prettyLog.append(tempFileName, " : The file was deleted from the server after the storage period.");
-					deleteCnt++;
+		if(fileList != null){
+			
+			for(File tempFile : fileList) {
+				if(tempFile.isFile()) {
+					String tempFileName = tempFile.getName();
+					prettyLog.append("FileName=", tempFileName);
+					
+					int extIdx = tempFileName.lastIndexOf(".");
+					String _tempFileName = tempFileName.substring(0, extIdx);
+					String __tempFileName = _tempFileName.replaceAll(substringTxt, "");
+					
+					int logDate_yyyy = Integer.parseInt(__tempFileName.substring(0, 4));
+					int logDate_MM = Integer.parseInt(__tempFileName.substring(4, 6));
+					int logDate_dd = Integer.parseInt(__tempFileName.substring(6, 8));
+					
+					Calendar cal = Calendar.getInstance();
+					cal.set(logDate_yyyy, logDate_MM-1, logDate_dd);
+					cal.set(Calendar.MILLISECOND, 0);
+					Date logDate = cal.getTime();
+					
+					long calDate = today.getTime()-logDate.getTime();
+					long calDateDays = calDate / ( 24*60*60*1000 ); 
+					
+					if(calDateDays > 31) {
+						tempFile.delete();
+						prettyLog.append(tempFileName, " : The file was deleted from the server after the storage period.");
+						deleteCnt++;
+					}
+					
 				}
-				
 			}
 		}
-		
-		for(File tempFile : fileList) {
-			if(tempFile.isFile()) {
-				String tempFileName = tempFile.getName();
-				prettyLog.append("final FileName=", tempFileName);
-				
+		if(fileList !=null){
+			
+			for(File tempFile : fileList) {
+				if(tempFile.isFile()) {
+					String tempFileName = tempFile.getName();
+					prettyLog.append("final FileName=", tempFileName);
+					
+				}
 			}
 		}
